Add change log in file NEWS
This file is called NEWS because CHANGELOG or ChangeLog assumes GNU
format, but in this file the format from

  http://keepachangelog.com/en/0.3.0/

is used.

Short description:

For every version of the program you use a block

[version] - date in YYYY-MM-DD

Added        --  for new features.
Changed      --  for changes in existing functionality.
Deprecated   --  for once-stable features removed in upcoming releases.
Removed for  --  deprecated features removed in this release.
Fixed        --  for any bug fixes.
Security     --  to invite users to upgrade in case of vulnerabilities.

Example:

[1.0.1] - 2016-01-21

Added
- Some feature6

Changed
- Some feature7

Fixed
- Some bug3

[1.0.0] - 2016-01-20

Added
- Some feature1
- Some feature2
- Some feature3

Changed
- Some feature4
- Some feature5

Fixed
- Some bug1
- Some bug2

End of example.